Sorry, forgot merges in "Blackboard"...
More code on the new implementation of a interview management in the d3web core. The interview management is "form-based" by now. So, you can ask session and interview management, respectively, to give the next "form" to be presented in a dialog. The "one question per form" strategy is already implemented.