Done some major refactoring in the d3web kernel. The ChoiceValue and MultipleChoiceValue are now no longer have references to any Choice object. This was required to decouple the values from a specific knowledge base instance. Instead, they have ChoiceID instances, that symbolically references the choices (that can be dereferenced to choices by handing over a specific question.
I've tried to test everything, seems to still work. Refactored but mostly untested is the imagequestion due to lack of knowledge from my side. The mc-questions in quicki still seems to work but sometimes showing somewhat unpredictable behavior. I think this was also the behavior before, because I found no wrong dereferencing of choices in my tests.

If there are any new problems with choice questions, please inform me to get this fixed!