1) Added headless wiki checker to verify html validity. For this the "valudator.nu" html parser has been integrated, which is also used in firefox browser. The test fails if any errors or warnings are produced by this parser.
2) Corrected all rendering bugs reported by the new html checker to fail for the headless testing wiki. The most common failures had been
* falsely closed <img> and <link> tags
* duplicate id's in html dom tree
* missing spaces between html attributes
* a few missing closing tags
* lots of false/missing encodings of "<", ">" when rendering keywords and terms
* many missing encodings of ' and " before they are placed inside html attributes

Due to lot's of changes (a few hundred code places in almost 100 files) it is expected that some errors will have been newly introduced, mainly e.g. in ajax actions or tool-rendering, because the functionality of js is not tested automatically by the html parser. I've tried to test as much as possible manually, but I'm almost sure I missed some things. Sorry for that in advance.