Use grouping only for large numbers >= 1,000,000 in num value verbalization
For smaller number, grouping doesn't seem like a good default, because the numbers are still easily readable without grouping and there is a chance where the grouping separator in english is misinterpreted as the decimal separator in german and vice versa. Also, in case the num value represents a year, the grouping also doesn't work (2,020 instead of 2020).