ILME-423: Edit getting test steps for verify repair of test equipment malfunctions
Add CostBenefitProperty to add malfunction IDs to a test step which should be performed for verifying these malfunctions.
Edit getting verify test steps in the mate: No need for checking all questions, it is enough to got through the test steps. But for backwards compatibility also questions are checked if no test step matches.